requirements.txt
aiogram
psycopg2-binary
requests
selenium
beautifulsoup4
fake-headers